feat(routine): extract pure domain into @sergeant/routine-domain (Phase 5 / PR 2)
Create new workspace package @sergeant/routine-domain mirroring the
conventions of @sergeant/finyk-domain: granular package.json exports,
standalone vitest suite (node env), no DOM-touching imports.

Extracted modules (100% pure, framework-agnostic):
  * types — Habit / Tag / Category / RoutineState / HabitDraft / etc.
  * constants — ROUTINE_TIME_MODES, RECURRENCE_OPTIONS, WEEKDAY_LABELS
  * dateKeys — dateKeyFromDate / parseDateKey / enumerateDateKeys /
    addDays / startOfIsoWeek / isoWeekdayFromDateKey
  * schedule — habitScheduledOnDate(habit, dateKey) predicate covering
    daily / weekdays / weekly / monthly / once recurrence with date
    bounds + archived flag
  * streaks — streakForHabit / maxStreakAllTime / maxActiveStreak /
    completionRateForRange / habitCompletionRate
  * drafts — routineTodayDate / REMINDER_PRESETS / normalizeReminderTimes /
    emptyHabitDraft / habitDraftToPatch
  * habitOrder — sortHabitsByOrder<T>(habits, order)
  * completionNoteKey

Storage-bound helpers stay in apps/web for now and will be split once
the matching MMKV shim lands in apps/mobile (Phase 7):
  * routineStorage (localStorage reads / writes)
  * hubCalendarAggregate.{loadMonthlyPlanDays,loadTemplateNameById,
    buildHubCalendarEvents} (safeReadLS)
  * finykSubscriptionCalendar (safeReadLS)

To preserve call-site stability, apps/web/src/modules/routine/lib/*.ts
re-export the pure helpers under their original import paths so no
downstream file needed updates beyond the files in this diff. The
Tailwind-coupled ROUTINE_THEME stays in web.

Tests: 6 new vitest suites in packages/routine-domain (34 assertions,
node env); redundant web copies removed; buildHubCalendarEvents test
narrowed to the storage-bound case.

Part of docs/react-native-migration.md §2.4 (Phase 5 / PR 2).
