+++ apps/web/src/modules/routine/lib/completionNoteKey.ts
@@ -1,3 +1,9 @@
-export function completionNoteKey(habitId, dateKey) {
-  return `${habitId}__${dateKey}`;
-}
+/**
+ * Completion-note composite key.
+ *
+ * Re-exports from the DOM-free `@sergeant/routine-domain` package
+ * (Phase 5 / PR 2). Historical import path is preserved so existing
+ * web call-sites don't have to change.
+ */
+
+export { completionNoteKey } from "@sergeant/routine-domain";
