+++ packages/routine-domain/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "@sergeant/routine-domain",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./constants": "./src/constants.ts",
+    "./date-keys": "./src/dateKeys.ts",
+    "./drafts": "./src/drafts.ts",
+    "./habit-order": "./src/habitOrder.ts",
+    "./schedule": "./src/schedule.ts",
+    "./streaks": "./src/streaks.ts",
+    "./types": "./src/types.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc -p tsconfig.json --noEmit",
+    "lint": "eslint \"src/**/*.{ts,tsx}\"",
+    "test": "vitest run",
+    "test:watch": "vitest"
+  },
+  "dependencies": {
+    "@sergeant/shared": "workspace:*"
+  },
+  "devDependencies": {
+    "@sergeant/config": "workspace:*",
+    "@types/node": "^25.6.0",
+    "typescript": "^6.0.3",
+    "vitest": "^3.0.0"
+  }
+}
