+++ packages/routine-domain/src/habitOrder.ts
@@ -1,0 +1,18 @@
+/** Стабільне сортування звичок за збереженим порядком */
+
+import type { Habit } from "./types.js";
+
+export function sortHabitsByOrder<T extends Habit = Habit>(
+  habits: T[],
+  order: string[] | null | undefined,
+): T[] {
+  const idx = new Map<string, number>(
+    (order || []).map((id, i) => [id, i] as const),
+  );
+  return [...habits].sort((a, b) => {
+    const ia = idx.has(a.id) ? (idx.get(a.id) as number) : 99999;
+    const ib = idx.has(b.id) ? (idx.get(b.id) as number) : 99999;
+    if (ia !== ib) return ia - ib;
+    return (a.name || "").localeCompare(b.name || "", "uk");
+  });
+}
