+++ packages/routine-domain/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "@sergeant/config/tsconfig.base.json",
+  "compilerOptions": {
+    "lib": ["ES2022"],
+    "types": ["node"],
+    "rootDir": "./src",
+    "outDir": "./dist"
+  },
+  "include": ["src/**/*"],
+  "exclude": ["node_modules", "dist"]
+}
